+++ rust/dbn-macros/tests/ui/csv_serialize_invalid_dbn_attr.rs
@@ -1,0 +1,10 @@
+use dbn_macros::CsvSerialize;
+
+#[derive(CsvSerialize)]
+#[repr(C)]
+struct Record {
+    #[dbn(unknown)]
+    pub a: u8,
+}
+
+fn main() {}