+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.7.0 - TBD
+## 0.7.1 - 2023-06-26
+- Added Python `variants` method to return an iterator over the enum variants for
+  `Compression`, `Encoding`, `Schema`, and `SType`
+- Improved Python enum conversions for `Compression`, `Encoding`, `Schema`, and `SType`
+
+## 0.7.0 - 2023-06-20
 - Added publishers enums
 - Added export to Python for `Compression`, `Encoding`, `SType`, and `Schema`
   enums